+++ CHANGES.md
@@ -51,6 +51,9 @@
   * Added `withUnfocused` function to `XMonad.Operations`, allowing for `X`
     operations to be applied to unfocused windows.
 
+  * Added `willFloat` function to `XMonad.ManageHooks` to detect whether the
+    (about to be) managed window will be a floating window or not
+
 [these build scripts]: https://github.com/xmonad/xmonad-testing/tree/master/build-scripts
 
 ### Bug Fixes